+++ TabletopTweaks-Base/MechanicsChanges/NaturalArmorStacking.cs
@@ -104,6 +104,7 @@ namespace TabletopTweaks.Base.MechanicsChanges {
                         .ForEach(c => c.Descriptor = (ModifierDescriptor)NaturalArmor.Stackable);
                     TTTContext.Logger.LogPatch("Patched", AnimalCompanionStatFeature);
                     AnimalCompanionUpgrades.ForEach(bp => {
+                        bp.SetName(TTTContext, "Animal Companion Upgrade");
                         bp.GetComponents<AddStatBonus>()
                             .Where(c => c.Descriptor == ModifierDescriptor.NaturalArmor)
                             .ForEach(c => c.Descriptor = (ModifierDescriptor)NaturalArmor.Stackable);
@@ -198,7 +199,7 @@ namespace TabletopTweaks.Base.MechanicsChanges {
                         .ForEach(c => c.Descriptor = (ModifierDescriptor)NaturalArmor.Size);
                     TTTContext.Logger.LogPatch("Patched", LegendaryProportionsBuff);
                     AnimalGrowthBuff.GetComponents<AddContextStatBonus>()
-                        .Where(c => c.Descriptor == ModifierDescriptor.NaturalArmorEnhancement)
+                        .Where(c => c.Descriptor == ModifierDescriptor.NaturalArmor)
                         .ForEach(c => c.Descriptor = (ModifierDescriptor)NaturalArmor.Stackable);
                 }
             }